Fixed a few things
So for mobile it doesn't register onmouseup so I had to make an ontouchup function so it would work properly on mobile. Aside from that I also removed webgl 2.0 as it may have been causing issues on apple mobile. So hopefully this fixes all my issues and doesn't cause any others.